+++ src/builtins.ts
@@ -1,8 +1,6 @@
-// @flow strict
-
 import { first } from './custom';
 import { count, ifilter, imap, izip, izip3, takewhile } from './itertools';
-import type { Maybe, Predicate, Primitive } from './types';
+import type { Predicate, Primitive } from './types';
 import { identityPredicate, keyToCmp, numberIdentity, primitiveIdentity } from './utils';
 
 /**
@@ -24,7 +22,7 @@ import { identityPredicate, keyToCmp, numberIdentity, primitiveIdentity } from '
  *
  */
 export function all<T>(iterable: Iterable<T>, keyFn: Predicate<T> = identityPredicate): boolean {
-    for (let item of iterable) {
+    for (const item of iterable) {
         if (!keyFn(item)) {
             return false;
         }
@@ -51,7 +49,7 @@ export function all<T>(iterable: Iterable<T>, keyFn: Predicate<T> = identityPred
  *
  */
 export function any<T>(iterable: Iterable<T>, keyFn: Predicate<T> = identityPredicate): boolean {
-    for (let item of iterable) {
+    for (const item of iterable) {
         if (keyFn(item)) {
             return true;
         }
@@ -88,9 +86,9 @@ export function contains<T>(haystack: Iterable<T>, needle: T): boolean {
  *     console.log([...enumerate(['hello', 'world'])]);
  *     // [0, 'hello'], [1, 'world']]
  */
-export function* enumerate<T>(iterable: Iterable<T>, start: number = 0): Iterable<[number, T]> {
+export function* enumerate<T>(iterable: Iterable<T>, start = 0): Iterable<[number, T]> {
     let index: number = start;
-    for (let value of iterable) {
+    for (const value of iterable) {
         yield [index++, value];
     }
 }
@@ -98,7 +96,7 @@ export function* enumerate<T>(iterable: Iterable<T>, start: number = 0): Iterabl
 /**
  * Non-lazy version of ifilter().
  */
-export function filter<T>(iterable: Iterable<T>, predicate: Predicate<T>): Array<T> {
+export function filter<T>(iterable: Iterable<T>, predicate: Predicate<T>): T[] {
     return Array.from(ifilter(iterable, predicate));
 }
 
@@ -108,17 +106,29 @@ export function filter<T>(iterable: Iterable<T>, predicate: Predicate<T>): Array
  * main use case of this function is to get a single iterator (a thing with
  * state, think of it as a "cursor") which can only be consumed once.
  */
-export function iter<T>(iterable: Iterable<T>): Iterator<T> {
-    // TODO: Not sure why Flow choked on this expression below, but at least we lock down the
-    // type transformation in the function signature this way.
-    // $FlowFixMe[incompatible-use]
-    return iterable[Symbol.iterator]();
+export function iter<T>(iterable: Iterable<T>): IterableIterator<T> {
+    // class SelfIter implements IterableIterator<T> {
+    //     #iterator: Iterator<T>;
+    //     constructor(orig: Iterable<T>) {
+    //         this.#iterator = orig[Symbol.iterator]();
+    //     }
+    //     [Symbol.iterator]() {
+    //         return this;
+    //     }
+    //     next() {
+    //         return this.#iterator.next();
+    //     }
+    // }
+    // return new SelfIter(iterable);
+
+    return iterable[Symbol.iterator]() as IterableIterator<T>;
+    //                                 ^^^^^^^^^^^^^^^^^^^^^^ Not safe!
 }
 
 /**
  * Non-lazy version of imap().
  */
-export function map<T, V>(iterable: Iterable<T>, mapper: (T) => V): Array<V> {
+export function map<T, V>(iterable: Iterable<T>, mapper: (item: T) => V): V[] {
     return Array.from(imap(iterable, mapper));
 }
 
@@ -133,7 +143,7 @@ export function map<T, V>(iterable: Iterable<T>, mapper: (T) => V): Array<V> {
  * If multiple items are maximal, the function returns either one of them, but
  * which one is not defined.
  */
-export function max<T>(iterable: Iterable<T>, keyFn: (T) => number = numberIdentity): Maybe<T> {
+export function max<T>(iterable: Iterable<T>, keyFn: (item: T) => number = numberIdentity): T | undefined {
     return reduce_(iterable, (x, y) => (keyFn(x) > keyFn(y) ? x : y));
 }
 
@@ -148,7 +158,7 @@ export function max<T>(iterable: Iterable<T>, keyFn: (T) => number = numberIdent
  * If multiple items are minimal, the function returns either one of them, but
  * which one is not defined.
  */
-export function min<T>(iterable: Iterable<T>, keyFn: (T) => number = numberIdentity): Maybe<T> {
+export function min<T>(iterable: Iterable<T>, keyFn: (item: T) => number = numberIdentity): T | undefined {
     return reduce_(iterable, (x, y) => (keyFn(x) < keyFn(y) ? x : y));
 }
 
@@ -157,7 +167,7 @@ export function min<T>(iterable: Iterable<T>, keyFn: (T) => number = numberIdent
  */
 function _range(start: number, stop: number, step: number): Iterable<number> {
     const counter = count(start, step);
-    const pred = step >= 0 ? (n) => n < stop : (n) => n > stop;
+    const pred = step >= 0 ? (n: number) => n < stop : (n: number) => n > stop;
     return takewhile(counter, pred);
 }
 
@@ -185,7 +195,7 @@ function _range(start: number, stop: number, step: number): Iterable<number> {
  * The produced range will be empty if the first value to produce already does
  * not meet the value constraint.
  */
-export function range(a: number, ...rest: Array<number>): Iterable<number> {
+export function range(a: number, ...rest: number[]): Iterable<number> {
     const args = [a, ...rest]; // "a" was only used by Flow to make at least one value mandatory
     switch (args.length) {
         case 1:
@@ -194,8 +204,9 @@ export function range(a: number, ...rest: Array<number>): Iterable<number> {
             return _range(args[0], args[1], 1);
         case 3:
             return _range(args[0], args[1], args[2]);
-        /* istanbul ignore next */
+
         default:
+            /* c8 ignore next */
             throw new Error('invalid number of arguments');
     }
 }
@@ -221,7 +232,7 @@ export function range(a: number, ...rest: Array<number>): Iterable<number> {
  * `reduce_()`, and the given iterable is empty, then no default value can be
  * derived and `undefined` will be returned.
  */
-export function reduce<T, O>(iterable: Iterable<T>, reducer: (O, T, number) => O, start: O): O {
+export function reduce<T, O>(iterable: Iterable<T>, reducer: (agg: O, item: T, index: number) => O, start: O): O {
     const it = iter(iterable);
     let output = start;
     for (const [index, item] of enumerate(it)) {
@@ -233,7 +244,7 @@ export function reduce<T, O>(iterable: Iterable<T>, reducer: (O, T, number) => O
 /**
  * See reduce().
  */
-export function reduce_<T>(iterable: Iterable<T>, reducer: (T, T, number) => T): Maybe<T> {
+export function reduce_<T>(iterable: Iterable<T>, reducer: (agg: T, item: T, index: number) => T): T | undefined {
     const it = iter(iterable);
     const start = first(it);
     if (start === undefined) {
@@ -258,9 +269,9 @@ export function reduce_<T>(iterable: Iterable<T>, reducer: (T, T, number) => T):
  */
 export function sorted<T>(
     iterable: Iterable<T>,
-    keyFn: (T) => Primitive = primitiveIdentity,
-    reverse: boolean = false
-): Array<T> {
+    keyFn: (item: T) => Primitive = primitiveIdentity,
+    reverse = false
+): T[] {
     const result = Array.from(iterable);
     result.sort(keyToCmp(keyFn)); // sort in-place