+++ src/index.ts
@@ -1,6 +1,4 @@
-// @flow strict
-
-import type { Predicate, Primitive } from './types';
+export type { Predicate, Primitive } from './types';
 
 export {
     all,
@@ -53,5 +51,3 @@ export {
     uniqueJustseen,
 } from './more-itertools';
 export { compact, compactObject, first, flatmap, icompact } from './custom';
-
-export type { Predicate, Primitive };