+++ src/types.js
@@ -1,5 +1,0 @@
-// @flow strict
-
-export type Maybe<T> = T | void;
-export type Predicate<T> = (T) => boolean;
-export type Primitive = string | number | boolean;