+++ src/types.ts
@@ -1,3 +1,2 @@
-export type Maybe<T> = T | undefined;
 export type Predicate<T> = (value: T) => boolean;
 export type Primitive = string | number | boolean;